Function pact_mock_server::lookup_mock_server

pub fn lookup_mock_server<R>(id: String, f: &Fn(&MockServer) -> R) -> Option<R>

Looks up the mock server by ID, and passes it into the given closure. The result of the closure is returned wrapped in an Option. If no mock server is found with that ID, None is returned.