Enum pact_mock_server::MatchResult

pub enum MatchResult {
    RequestMatch(Interaction),
    RequestMismatch(Interaction, Vec<Mismatch>),
    RequestNotFound(Request),
    MissingRequest(Interaction),
}

Enum to define a match result

Variants
RequestMatch(Interaction)

Match result where the request was sucessfully matched

RequestMismatch(Interaction, Vec<Mismatch>)

Match result where there were a number of mismatches

RequestNotFound(Request)

Match result where the request was not expected

MissingRequest(Interaction)

Match result where an expected request was not received

Methods

impl MatchResult

fn match_key(&self) -> String

Returns the match key for this mismatch

fn matched(&self) -> bool

Returns true if this match result is a RequestMatch

fn to_json(&self) -> Json

Converts this match result to a Json struct